% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspace_alignment_Fernando2013.R
\name{alignSubspace}
\alias{alignSubspace}
\title{Subspace alignment}
\usage{
alignSubspace(sourceFeatures, targetFeatures, d = NULL, sigma = 0.1,
  gamma = 10^5, B = 1)
}
\arguments{
\item{sourceFeatures}{source dataset}

\item{targetFeatures}{target dataset}

\item{d}{dimension of the subspace}

\item{sigma}{related to theoratical lower bound. 0.1 by default}

\item{gamma}{related to theoratical lower bound. 10^5 by default}

\item{B}{related to theoratical lower bound. 1 by default}
}
\value{
list object containing transformed (aligned) source and target data, d, lower bound, and differences in eigen values
}
\description{
Subspace alignment
}
