% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfManipulation.R
\name{cyname}
\alias{cyname}
\title{Rename a Response Column}
\usage{
cyname(df, orgname, altname = "y")
}
\arguments{
\item{df}{A data frame.}

\item{orgname}{Character string specifying the original column name.}

\item{altname}{Character string specifying the new column name.}
}
\value{
A data frame with the renamed column.
}
\description{
Change the name of a specified column in a data frame.
}
\examples{
df <- data.frame(x = 1:3, y = 4:6)
cyname(df, "y", "response")

}
\seealso{
Other dataframe_tools: 
\code{\link{factorizevars}()},
\code{\link{na_count}()},
\code{\link{na_eachcol}()},
\code{\link{setrownames}()},
\code{\link{siftcoltype}()}
}
\concept{dataframe_tools}
