% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{dateToIsoWeek}
\alias{dateToIsoWeek}
\title{Convert an iso date to simple isoweek format}
\usage{
dateToIsoWeek(dates)
}
\arguments{
\item{dates}{ISOdates as string vector}
}
\value{
a vector of strings with ISOweeks inside
}
\description{
Simplistic representation of an isoweek format is encoded here. Basically we convert 2017-11-23 to 201747 thus YYYY-MM-DD to YYYYWW.
}
\examples{
dateToIsoWeek(c("2017-11-23", "2015-11-23"))
}
