% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datools.R
\docType{package}
\name{datools}
\alias{datools}
\alias{datools-package}
\title{datools: A set of useful tools for machine learning consulting using R.}
\description{
The datools package provides a set of useful tools for machine learning
consulting and other data science oriented solutions using R.
}
\section{Datools functions}{

Well there's a lot of them.
}

