% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablerelationships.R
\name{discover_hierarchy_and_fit}
\alias{discover_hierarchy_and_fit}
\title{Find the relationship between the variables in a data set and fit that model}
\usage{
discover_hierarchy_and_fit(data)
}
\arguments{
\item{data}{the data.frame containing the observations and variables to inspect}
}
\value{
a fitted architecture
}
\description{
Finds the hirarchical relationship between all variables and connects them in a network such as
to optimize the mutual information. This just fits the network selected by HillClimbing using a
Bayesian information criteria (BIC).
}
\examples{
data(mtcars)
library(dplyr)
myfit <- discover_hierarchy_and_fit(mtcars)
if(requireNamespace("Rgraphviz")) bnlearn::graphviz.plot(myfit)
}
