% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variablerelationships.R
\name{discover_variable_hierarchies}
\alias{discover_variable_hierarchies}
\title{Find the relationship between the variables in a data set}
\usage{
discover_variable_hierarchies(data, n = 5)
}
\arguments{
\item{data}{the data.frame containing the observations and variables to inspect}

\item{n}{the number of networks to consider}
}
\value{
a character matrix with columns from and to representing the variables that are related
}
\description{
Finds the hirarchical relationship between all variables and connects them in a network such as
to optimize the mutual information.
}
\examples{
data(iris)
data(mtcars)
library(dplyr)
discover_variable_hierarchies(mtcars, 10) \%>\% as_tibble()
discover_variable_hierarchies(iris, 10) \%>\% as_tibble()
}
