% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Draw an Empty Plot}
\usage{
emptyPlot(text = "")
}
\arguments{
\item{text}{Character string to display in the center of the plot.
If an empty string (default), no text is drawn.}
}
\value{
A recorded plot object produced by \code{recordPlot()}.
}
\description{
Create an empty plot that can be used as a placeholder when arranging
multiple figures in a grid layout. Optionally, a text label can be
displayed in the center of the plot.
}
\examples{
emptyPlot()
emptyPlot("No Data")

}
\seealso{
Other plot_tools: 
\code{\link{generateColors}()}
}
\concept{plot_tools}
