% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunctions.R
\name{extrapolateNA}
\alias{extrapolateNA}
\title{Extrapolate NA's in a univariate vector}
\usage{
extrapolateNA(x, len = 3)
}
\arguments{
\item{x}{the univariate vector to extrapolate}

\item{len}{the number of non NA elements to base the extrapolation on which
defaults to 3}
}
\value{
a vector with the NA's in the beginning and the end extrapolated
}
\description{
Extrapolates NA's in the beginning or the end of a univariate vector.
Intermediate NA's that occur between two observed values are not touched.
Please use interpolation for that case.
}
\examples{
extrapolateNA(c(NA, NA, NA, 4, 5, 6, 7, NA, NA))
extrapolateNA(c(0, NA, NA, 4, 5, 6, 7, NA, NA))
extrapolateNA(c(0, NA, NA, 4, 5, 6, 7, NA, 9))
}
