% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfManipulation.R
\name{factorizevars}
\alias{factorizevars}
\title{Convert Variables to Factors}
\usage{
factorizevars(df, otherTypes = c(), specificCols = c())
}
\arguments{
\item{df}{A data frame.}

\item{otherTypes}{Character vector specifying additional data types
that should be converted to factors.}

\item{specificCols}{Character vector specifying column names that
should be converted to factors regardless of their type.}
}
\value{
A data frame with specified variables converted to factors.
}
\description{
Convert selected columns of a data frame into factors.
By default, columns of type \code{character} and \code{logical} are converted.
Additional column types or specific columns can also be specified.
}
\examples{
df <- data.frame(
  a = letters[1:3],
  b = c(TRUE, FALSE, TRUE),
  c = 1:3
)

factorizevars(df)

}
\seealso{
Other dataframe_tools: 
\code{\link{cyname}()},
\code{\link{na_count}()},
\code{\link{na_eachcol}()},
\code{\link{setrownames}()},
\code{\link{siftcoltype}()}
}
\concept{dataframe_tools}
