% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{find_lag}
\alias{find_lag}
\title{Find the optimal lag}
\usage{
find_lag(x, y, lag.max = as.integer(length(x)/2))
}
\arguments{
\item{x}{time series x}

\item{y}{time series y}

\item{lag.max}{the maximum number of lags to investigate which defaults to
the integer value of length(x)/2.}
}
\value{
an integer representign the optimal lag to apply to x
}
\description{
Finds the lag that maximizes the absolute correlation between two time
series x and y.
}
\examples{
library(tibble)
library(ggplot2)
library(tidyr)
library(dplyr)
x <- sin(seq(1, 10, length.out = 30))
y <- lag(x,5) + rnorm(length(x), 0, 0.2)
testdf <- tibble(id=seq_along(x), x, y)
ggplot(gather(testdf, key, val, -id), aes(y=val, x=id, color=key)) + geom_line()
stats::ccf(x,y, na.action = na.omit, lag.max = 5)
testdf <- mutate(testdf, z=lag(x, find_lag(x,y)))
ggplot(gather(testdf, key, val, -id), aes(y=val, x=id, color=key)) + geom_line()
}
