% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generateColors}
\alias{generateColors}
\title{Generate Distinct Colors}
\usage{
generateColors(number)
}
\arguments{
\item{number}{Integer specifying the number of colors to generate.}
}
\value{
A character vector containing hexadecimal color codes.
}
\description{
Generate a set of visually distinct colors using evenly spaced values
in the HSV color space.
}
\examples{
generateColors(5)

}
\seealso{
Other plot_tools: 
\code{\link{emptyPlot}()}
}
\concept{plot_tools}
