% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{genHolidays}
\alias{genHolidays}
\title{Generate public holidays for all markets}
\usage{
genHolidays(
  years = (lubridate::year(Sys.Date()) - 2):lubridate::year(Sys.Date()),
  country
)
}
\arguments{
\item{years}{the years for which to generate the holidays given as 2013:2017
as an example where the default is the last three years}

\item{country}{specifies which specific country to focus on except the "global"
holidays where the possible values are "CA", "CH", "DE", "FR", "GB", "IT", "JP", "US"}
}
\value{
a one hot encoded tibble where each column is a holiday and the first
column is the ISO date
}
\description{
The function heavily relies on the timeDate package in R for all holidays.
Specifically it utilizes the listHolidays function to figure out which holidays
are existing.
}
\examples{
b <- genHolidays(2016:2017)
b[1:10, 1:5]
}
