% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunctions.R
\name{interpolateNA}
\alias{interpolateNA}
\title{Perform linear interpolation of a time series}
\usage{
interpolateNA(x)
}
\arguments{
\item{x}{the vector to interpolate}
}
\value{
an interpolated vector
}
\description{
Just what it says with linear interpolation.
}
\examples{
interpolateNA(c(1,2,3, NA, NA, NA, 7, 8, 9))
}
