% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableencoding.R
\name{lagvariables}
\alias{lagvariables}
\title{Lag variables in data nlags}
\usage{
lagvariables(x, nlag = 5)
}
\arguments{
\item{x}{the data tibble to operate on}

\item{nlag}{the max number of lags to create for each variable}
}
\value{
a new tibble with the data lagged
}
\description{
This function uses an existing dataset in the shape of a tibble and then
lags each variable(column) in the tibble from 1 all the way up to nlag.
}
\details{
This function requires that the first column in your tibble is called
"date" and represents some kind of time epoch. Note also that you will
get no NA's returned. This means that the returned data will start nlag-1
days later than the input data.
}
\examples{
library(dplyr)
tmpdf <- tibble(date=seq(from=as.Date("2018-01-01"), length.out=10, by="1 day"),
                a=1:10, b=11:20, d=21:30)
lagdf <- tmpdf \%>\% lagvariables(4)
lagdf
tmpdf
}
