% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(o, p)
}
\arguments{
\item{o}{the observed value (trueth)}

\item{p}{the predicted value (estimation)}
}
\value{
a scalar value
}
\description{
Mean absolute error
}
