% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mase}
\alias{mase}
\title{Mean absolute scaled error}
\usage{
mase(o, p)
}
\arguments{
\item{o}{the observed value (trueth)}

\item{p}{the predicted value (estimation)}
}
\value{
a scalar value
}
\description{
Mean absolute scaled error
}
