% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(o, p)
}
\arguments{
\item{o}{the observed value (trueth)}

\item{p}{the predicted value (estimation)}
}
\value{
a scalar value
}
\description{
Mean squared error
}
