% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfManipulation.R
\name{na_count}
\alias{na_count}
\title{Count Missing Values}
\usage{
na_count(vec)
}
\arguments{
\item{vec}{A vector.}
}
\value{
Integer count of \code{NA} values.
}
\description{
Count the number of missing (\code{NA}) values in a vector.
}
\examples{
na_count(c(1, NA, 3, NA))

}
\seealso{
Other dataframe_tools: 
\code{\link{cyname}()},
\code{\link{factorizevars}()},
\code{\link{na_eachcol}()},
\code{\link{setrownames}()},
\code{\link{siftcoltype}()}
}
\concept{dataframe_tools}
