% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfManipulation.R
\name{na_eachcol}
\alias{na_eachcol}
\title{Count Missing Values for Each Column}
\usage{
na_eachcol(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A named numeric vector where each element represents
the number of \code{NA} values in the corresponding column.
}
\description{
Compute the number of missing (\code{NA}) values in each column
of a data frame.
}
\examples{
df <- data.frame(
  a = c(1, NA, 3),
  b = c(NA, NA, 2)
)

na_eachcol(df)

}
\seealso{
Other dataframe_tools: 
\code{\link{cyname}()},
\code{\link{factorizevars}()},
\code{\link{na_count}()},
\code{\link{setrownames}()},
\code{\link{siftcoltype}()}
}
\concept{dataframe_tools}
