% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunctions.R
\name{naToVal}
\alias{naToVal}
\title{Replace NA with a user given value in a vector}
\usage{
naToVal(x, y = 0)
}
\arguments{
\item{x}{the vector to replace all NA in}

\item{y}{the value to replace the NA with}
}
\value{
a new vector with all NA's replaced by y
}
\description{
This function replaces all NA's in a vector with a value provided by the user.
}
\examples{
naToZero(c(NA, 3:5, NA, NA, 4, 4, 10, NA))
naToVal(c(NA, 3:5, NA, NA, 4, 4, 10, NA), 0)
}
