% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunctions.R
\name{naToZero}
\alias{naToZero}
\title{Replace NA with 0 in a vector}
\usage{
naToZero(x)
}
\arguments{
\item{x}{the vector to replace all NA in}
}
\value{
a new vector with all NA's replaced by 0
}
\description{
This function replaces all NA's in a vector with 0
}
\examples{
naToZero(c(NA, 3:5, NA, NA, 4, 4, 10, NA))
}
