% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuefunctions.R
\name{nonFinToVal}
\alias{nonFinToVal}
\title{Replace non finite values with a user given value in a vector}
\usage{
nonFinToVal(x, y = 0)
}
\arguments{
\item{x}{the vector to replace all non finite values in}

\item{y}{the value to replace the non finite values with}
}
\value{
a new vector with all non finite values replaced by y
}
\description{
This function replaces all non finite values in a vector with a value provided
by the user. NA, NaN, Inf are all examples of non finite values.
}
\examples{
nonFinToVal(c(NA, 3:5, NA, NA, 4, 4, 10, NA))
nonFinToVal(c(NA, 3:5, NA, NA, 4, 4, 10, NA), 3)
}
