% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascaling.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector of numbers}
\usage{
normalize(x)
}
\arguments{
\item{x}{the vector of numbers to normalize}
}
\value{
the normalized vector of numbers
}
\description{
This normalizes a vector of numbers. Normarlization rescales the original
data into a range with 0 mean and unit variance.
}
\examples{
normalize(rnorm(10, 20, 10))
}
