% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableencoding.R
\name{oneHotEncoder2}
\alias{oneHotEncoder2}
\title{Encode a vector into a one hot tibble}
\usage{
oneHotEncoder2(x, f = function(y) y)
}
\arguments{
\item{x}{the vector to convert to a one hot}

\item{f}{a function to apply to x before encoding which defaults to identity}
}
\value{
the one hot encoded tibble with the original data given in a Data column
}
\description{
This function takes a vector of items and onehot encodes them into a tibble
which in reality is a tibble. Every unique value of the vector get converted into
a column and the value is 1 for that column if the row number matches the position
in the original vector. See the examples for details.
}
\examples{
library(lubridate)
oneHotEncoder2(x=seq(as.Date("2017-01-01"), by ="day", length.out = 10),
              f=function(x) wday(x, label = TRUE))
oneHotEncoder2(x=wday(seq(as.Date("2017-10-07"), by ="days", length.out = 10), label = TRUE))
}
