% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Make a waterfall plot from a tibble of cash flows}
\usage{
plot_waterfall(data)
}
\arguments{
\item{data}{the data tibble to operate on}
}
\value{
a ggplot of the waterfall plot
}
\description{
Make a waterfall plot from a tibble of cash flows
}
\examples{
datadf <-
tibble::tibble(
  desc = c(
    'starting_net_income',
    'write_offs',
    'sales',
    'COGS',
    'overhead',
    'taxes',
    'bonuses',
    'other_income',
    'ending_net_income'
  ),
  amount = c(1200, -2300, 11000, -9000, -1900, -1200, -3000, 7000, 1800)
)
plot_waterfall(datadf)
}
