% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatools.R
\name{plotPCAComponent}
\alias{plotPCAComponent}
\title{Plot original data in PCA space}
\usage{
plotPCAComponent(mytibble, classes = NULL)
}
\arguments{
\item{mytibble}{the data frame to operate on which can only contain numeric values}

\item{classes}{optional class for each observation in the data frame}
}
\value{
a ggplot2 object representing the PCA plot
}
\description{
This visualizes the directionality of the original data in Principal
Component Analysis (PCA) space.
}
\examples{
library(datools)
plotPCAComponent(iris[,-5], iris$Species) + theme_minimal()
}
