% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Plot prediction based on a time series}
\usage{
plotPrediction(preddf)
}
\arguments{
\item{preddf}{the prediction data to plot}
}
\value{
a ggplot object
}
\description{
This plots estmated value vs observed for a given time series.
It also visualized upper and lower level for that estimate.
}
\examples{
library(ggplot2)
library(dplyr)
data(economics)
myfit <- lm(unemploy~pop+psavert, data=economics)
preddf <- predict(myfit, interval="predict") \%>\% as_tibble()
preddf <- preddf \%>\% transmute(date=as.Date(economics$date),
                               observed=economics$unemploy,
                               estimate=fit,
                               lower=lwr,
                               upper=upr)
plotPrediction(preddf)
}
