% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascaling.R
\name{rangeToBuckets}
\alias{rangeToBuckets}
\title{Split a vector of indices into buckets}
\usage{
rangeToBuckets(x, bucketsize = 10)
}
\arguments{
\item{x}{the indices to split into buckets}

\item{bucketsize}{the size each bucket should have}
}
\value{
a list of indices of length bucketsize
}
\description{
Splits any index vector into a list of index vectors where each vector is of
length bucketsize exept for the last one which contains the remainder. The
remainder is defined as length(x)%%bucketsize. It works with all kind of
discrete ranges. Negative to positive.
}
\examples{
rangeToBuckets(1:nrow(mtcars), 9)
rangeToBuckets(-2:3, 4)
}
