% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfManipulation.R
\name{setrownames}
\alias{setrownames}
\title{Set Row Names from a Column}
\usage{
setrownames(df, column, delete = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{column}{Character string specifying the column to use as row names.}

\item{delete}{Logical indicating whether the column should be removed
after assigning row names. Default is \code{TRUE}.}
}
\value{
A data frame with updated row names.
}
\description{
Use the values of a specified column as the row names of a data frame.
Optionally remove that column afterward.
}
\examples{
df <- data.frame(id = c("a","b","c"), value = 1:3)
setrownames(df, "id")

}
\seealso{
Other dataframe_tools: 
\code{\link{cyname}()},
\code{\link{factorizevars}()},
\code{\link{na_count}()},
\code{\link{na_eachcol}()},
\code{\link{siftcoltype}()}
}
\concept{dataframe_tools}
