% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascaling.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector of numbers}
\usage{
standardize(x, a = 0, b = 1)
}
\arguments{
\item{x}{the vector of numbers to standardize}

\item{a}{the lower limit of the wanted scale}

\item{b}{the upper limit of the wanted scale}
}
\value{
the rescaled vector of numbers
}
\description{
This standardizes a vector of numbers. Standardization rescales the original
data into a range between a and b which are given as parameters. They are 0
and 1 respectively by default.
}
\examples{
standardize(rnorm(10, 20, 10))
}
