% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmodels.R
\name{stripGlmModel}
\alias{stripGlmModel}
\title{Strip a fitted GLM model}
\usage{
stripGlmModel(a)
}
\arguments{
\item{a}{the glm model object to strip}
}
\value{
a stripped GLM model
}
\description{
Reduce the size of a fitted GLM model by a massive factor. This is crucial in
deploying R in production settings as the standard GLM is way to liberal with
storing stuff we don't need for prediction.
}
\examples{
mylm <- glm(dist ~ speed, data = cars)
print(object.size(mylm))
mylm <- stripGlmModel(mylm)
print(object.size(mylm))
print(predict(mylm, cars))
}
