% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCA_Pan2009.R
\name{tca}
\alias{tca}
\title{TCA}
\usage{
tca(sourceFeatures, targetFeatures, m, sigma = 1, mu = 0.1)
}
\arguments{
\item{sourceFeatures}{source data}

\item{targetFeatures}{target data}

\item{m}{number of components}

\item{sigma}{width of RBF kernel}

\item{mu}{trade-off parameter}
}
\description{
TCA
}
