% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsefunctions.R
\name{tf_dagum}
\alias{tf_dagum}
\title{The Dagum transformation function}
\usage{
tf_dagum(x, a = 1, b = 1, p = 1)
}
\arguments{
\item{x}{value to transform}

\item{a}{the scale parameter}

\item{b}{the first shape parameter}

\item{p}{the second shape parameter}
}
\value{
x transformed
}
\description{
An implementation of the CDF for the Dagum distribution.
}
\examples{
library(ggplot2)
ggplot(data.frame(x = c(0, 4)), aes(x)) +
  stat_function(fun = function(x) tf_dagum(x, 1, 2, 2))
}
