% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsefunctions.R
\name{tf_relunoise}
\alias{tf_relunoise}
\title{The Noise ReLu transformation function}
\usage{
tf_relunoise(x)
}
\arguments{
\item{x}{value to transform}
}
\value{
x transformed
}
\description{
An implementation of the Noisy ReLu transformation function
}
\examples{
library(ggplot2)
ggplot(data.frame(x = c(0, 4)), aes(x)) +
  stat_function(fun = function(x) tf_relunoise(x))
\dontrun{
library(microbenchmark)
microbenchmark::microbenchmark(tf_relu(1:100), tf_softplus(1:100),
  tf_dagum(1:100), tf_sigmoid(1:100),
  tf_fisk(1:100), tf_relunoise(1:100),
  times = 10000
)
}
}
