% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsefunctions.R
\name{tf_softplus}
\alias{tf_softplus}
\title{The Softplus ReLu transformation function}
\usage{
tf_softplus(x)
}
\arguments{
\item{x}{value to transform}
}
\value{
x transformed
}
\description{
An implementation of the Softplus ReLu transformation function
}
\examples{
library(ggplot2)
ggplot(data.frame(x = c(0, 4)), aes(x)) +
  stat_function(fun = function(x) tf_softplus(x))
}
