% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_dir}
\alias{time_dir}
\title{Format Time for Directory Names}
\usage{
time_dir(t)
}
\arguments{
\item{t}{A time object (typically \code{POSIXct}).}
}
\value{
A character string formatted as \code{"YYYY-MM-DD_HH-MM-SS"}.
}
\description{
Convert a time object into a string formatted for use in directory
or file names.
}
\examples{
time_dir(Sys.time())

}
\seealso{
Other time_tools: 
\code{\link{time_convert}()},
\code{\link{time_log}()}
}
\concept{time_tools}
