% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{to_waterfall}
\alias{to_waterfall}
\title{Make waterfall data tibble from a tibble of cash flows}
\usage{
to_waterfall(data)
}
\arguments{
\item{data}{the data tibble to operate on}
}
\value{
a new data tibble with columns necessary to create a waterfall visualization
}
\description{
The only requirement for this function to work is that the first row and last row of the tibble
represent the cash in and cash out. In other words the Base and the Total when talking about
sales modeling.
}
\examples{
datadf <-
  tibble::tibble(
    desc = c(
      "starting_net_income",
      "write_offs",
      "sales",
      "COGS",
      "overhead",
      "taxes",
      "bonuses",
      "other_income",
      "ending_net_income"
    ),
    amount = c(1200, -2300, 11000, -9000, -1900, -1200, -3000, 7000, 1800)
  )
to_waterfall(datadf)
# Change order
datadf2 <- to_waterfall(datadf)
firstindex <- which(datadf$desc == "sales")
datadf2[c(firstindex, setdiff(seq_along(datadf$amount), firstindex)), ]
}
