% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_args.R
\name{validate_input}
\alias{validate_input}
\title{Validate Inputs}
\usage{
validate_input(valid_inputs, inputs, all = "all")
}
\arguments{
\item{valid_inputs}{Character vector of allowed inputs.}

\item{inputs}{Character vector of user-specified inputs.}

\item{all}{Character string representing a keyword that expands to all
valid inputs. Default is \code{"all"}.}
}
\value{
A character vector containing validated inputs.
}
\description{
Validate a set of user-specified inputs against a list of allowed values.
If the keyword specified by \code{all} appears in \code{inputs}, all valid values
are returned.
}
\examples{
valid <- c("clean", "train", "predict")

validate_input(valid, c("clean", "train"))
validate_input(valid, "all")

}
\seealso{
Other cli_tools: 
\code{\link{get_all_cliargs}()},
\code{\link{parse_cliarg}()}
}
\concept{cli_tools}
