% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodelsutils.R
\name{vif.data.frame}
\alias{vif.data.frame}
\title{Variance inflation factors of a data frame}
\usage{
\method{vif}{data.frame}(mod, ...)
}
\arguments{
\item{mod}{A data.frame with numerical and factor variables}
}
\value{
A matrix containing the VIF and standardised VIF factors
}
\description{
Contains a wrapper function around car::vif to calculate the variance inflation of data.frames
}
\examples{
a <- 1
}
