% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodelsutils.R
\name{vif.lm}
\alias{vif.lm}
\title{Calculate the VIF for data}
\usage{
\method{vif}{lm}(mod)
}
\arguments{
\item{mod}{the data to process}
}
\value{
A matrix containing the VIF and standardised VIF factors
}
\description{
Calculate the Variance Inflation Factor for a dataset.
}
\examples{
a <- 1
}
