% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{wmape}
\alias{wmape}
\title{Weighted mean absolute percentage error}
\usage{
wmape(o, p)
}
\arguments{
\item{o}{the observed value (trueth)}

\item{p}{the predicted value (estimation)}
}
\value{
a scalar value
}
\description{
Weighted mean absolute percentage error
}
