% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunManager.R
\name{RunManager}
\alias{RunManager}
\title{RunManager}
\description{
Manages project IO paths and execution lifecycle.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RunManager-new}{\code{RunManager$new()}}
\item \href{#method-RunManager-c_path}{\code{RunManager$c_path()}}
\item \href{#method-RunManager-d.raw_path}{\code{RunManager$d.raw_path()}}
\item \href{#method-RunManager-d.usable_path}{\code{RunManager$d.usable_path()}}
\item \href{#method-RunManager-d.preped_path}{\code{RunManager$d.preped_path()}}
\item \href{#method-RunManager-o.runs_path}{\code{RunManager$o.runs_path()}}
\item \href{#method-RunManager-o.published_path}{\code{RunManager$o.published_path()}}
\item \href{#method-RunManager-s.R_path}{\code{RunManager$s.R_path()}}
\item \href{#method-RunManager-s.exe_path}{\code{RunManager$s.exe_path()}}
\item \href{#method-RunManager-actrun_path}{\code{RunManager$actrun_path()}}
\item \href{#method-RunManager-actruninfo_path}{\code{RunManager$actruninfo_path()}}
\item \href{#method-RunManager-consolelog_path}{\code{RunManager$consolelog_path()}}
\item \href{#method-RunManager-record_git_info}{\code{RunManager$record_git_info()}}
\item \href{#method-RunManager-snapshot_env}{\code{RunManager$snapshot_env()}}
\item \href{#method-RunManager-parse_cliarg}{\code{RunManager$parse_cliarg()}}
\item \href{#method-RunManager-validate_input}{\code{RunManager$validate_input()}}
\item \href{#method-RunManager-default_config}{\code{RunManager$default_config()}}
\item \href{#method-RunManager-source_scripts}{\code{RunManager$source_scripts()}}
\item \href{#method-RunManager-logmsg}{\code{RunManager$logmsg()}}
\item \href{#method-RunManager-log_system_info}{\code{RunManager$log_system_info()}}
\item \href{#method-RunManager-get_config}{\code{RunManager$get_config()}}
\item \href{#method-RunManager-get_cliargs}{\code{RunManager$get_cliargs()}}
\item \href{#method-RunManager-initialize_run}{\code{RunManager$initialize_run()}}
\item \href{#method-RunManager-run}{\code{RunManager$run()}}
\item \href{#method-RunManager-clone}{\code{RunManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$new(root)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-c_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-c_path}{}}}
\subsection{Method \code{c_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$c_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-d.raw_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-d.raw_path}{}}}
\subsection{Method \code{d.raw_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$d.raw_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-d.usable_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-d.usable_path}{}}}
\subsection{Method \code{d.usable_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$d.usable_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-d.preped_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-d.preped_path}{}}}
\subsection{Method \code{d.preped_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$d.preped_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-o.runs_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-o.runs_path}{}}}
\subsection{Method \code{o.runs_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$o.runs_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-o.published_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-o.published_path}{}}}
\subsection{Method \code{o.published_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$o.published_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-s.R_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-s.R_path}{}}}
\subsection{Method \code{s.R_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$s.R_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-s.exe_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-s.exe_path}{}}}
\subsection{Method \code{s.exe_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$s.exe_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-actrun_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-actrun_path}{}}}
\subsection{Method \code{actrun_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$actrun_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-actruninfo_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-actruninfo_path}{}}}
\subsection{Method \code{actruninfo_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$actruninfo_path(file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-consolelog_path"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-consolelog_path}{}}}
\subsection{Method \code{consolelog_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$consolelog_path(dir, file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-record_git_info"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-record_git_info}{}}}
\subsection{Method \code{record_git_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$record_git_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-snapshot_env"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-snapshot_env}{}}}
\subsection{Method \code{snapshot_env()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$snapshot_env()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-parse_cliarg"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-parse_cliarg}{}}}
\subsection{Method \code{parse_cliarg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$parse_cliarg(args, flag)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-validate_input"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-validate_input}{}}}
\subsection{Method \code{validate_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$validate_input(valid_args, args_input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-default_config"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-default_config}{}}}
\subsection{Method \code{default_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$default_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-source_scripts"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-source_scripts}{}}}
\subsection{Method \code{source_scripts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$source_scripts(scriptpath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-logmsg"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-logmsg}{}}}
\subsection{Method \code{logmsg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$logmsg(msg, level = "INFO")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-log_system_info"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-log_system_info}{}}}
\subsection{Method \code{log_system_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$log_system_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-get_config}{}}}
\subsection{Method \code{get_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$get_config(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-get_cliargs"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-get_cliargs}{}}}
\subsection{Method \code{get_cliargs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$get_cliargs(name = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-initialize_run"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-initialize_run}{}}}
\subsection{Method \code{initialize_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$initialize_run(
  libraries = character(0),
  scriptpath = character(0),
  flags = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-run"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$run(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RunManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
