% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodelsutils.R
\name{vif}
\alias{vif}
\title{Calculate the VIF for data}
\usage{
vif(mod)
}
\arguments{
\item{mod}{the data to process}
}
\value{
the VIF for the given data
}
\description{
Calculate the Variance Inflation Factor for a dataset.
}
\examples{
a <- 1
}
